chapters

+++ app/src/main/java/org/grakovne/lissen/playback/service/PlaybackService.kt
@@ -157,7 +157,7 @@ class AudioPlayerService : MediaSessionService() {
         when (position) {
             null -> exoPlayer.seekTo(0, 0)
             else -> {
-                val duration = chapters.runningFold(0.0) { acc, chapter -> acc + chapter.duration.toLong() }
+                val duration = chapters.runningFold(0.0) { acc, chapter -> acc + chapter.duration }
                 val targetChapter = duration.indexOfFirst { it > position }
                 val chapterProgress = position - duration[targetChapter - 1]
